Redefine/requirements.txt
requests~=2.28.2
ruamel.yaml~=0.18.6